// src/app/models/product.model.ts
export interface AddToCartReqObj {
    MobileNo: string;
    SKU: string;
    VendorSKU: string;
    VendorName: string;
    VendorCode: string;
    StoreCode: string;
    StoreName: string;
    ProductType: string;
    ProductName: string;
    Category: string;
    TaxAmount: number;
    TaxPercentage: number;
    MRP: number;
    Discount: number;
    Quantity: number;
    ImagePath: string;
    TotalMRP: number;
    TotalDiscount: number;
    PriceSaved: number;
    Rating: number;
    HSNCode: string;
    BrandCode?: string;
    BrandValue?: string;
    ColoursCode?: string;
    ColoursValue?: string;
    CoursesValue?: string;
    CoursesCode?: string;
    DevicetypesValue?: string;
    DevicetypesCode?: string;
    GenderValue?: string;
    GenderCode?: string;
    GradesValue?: string;
    GradesCode?: string;
    LanguagesCode?: string;
    LanguagesValue?: string;
    PublishersCode?: string;
    PublishersValue?: string;
    SizesCode?: string;
    SizesValue?: string;
    SubjectsCode?: string;
    SubjectsValue?: string;
    Variants?: any;
    ProductID?: string;
    Inventory?: number;
    MinimumBuyQty?: number;
    MaximumBuyQty?: number;
    IsSpecialProduct?: string;
    DetailsID?: string;
    BaseUnitPrice?:string;
	DiscountPrice?:number;
	DiscountValue?:number;
	MinimumPrice?:number;
	MaximumPrice?:number;
	GSTP?:number;
	GSTV?:number;
    SKUNature?:string;
}